+++ gen/com/kaylerrenslow/armaplugin/lang/sqf/psi/SQFPsiSignedExpression.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
 
-public interface SQFPsiUnaryExpression extends SQFPsiExpression {
+public interface SQFPsiSignedExpression extends SQFPsiExpression {
 
   @Nullable
   SQFPsiExpression getExpression();